backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
opencv-python-headless==4.11.0.86
numpy==2.2.6
Pillow==11.2.1
requests==2.32.3
python-multipart==0.0.20
scipy==1.15.3
